src/test/kotlin/Day20Test.kt
import org.eruanno.Day20
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.CsvSource

class Day20Test {
    private val subject = Day20()

    @ParameterizedTest
    @CsvSource(
        "'day-20-test', 44",
        "'day-20', 1289"
    )
    fun testPartOne(testFile: String, expectedResult: Int) {
        assertEquals(expectedResult, subject.partOne(testFile))
    }

    @ParameterizedTest
    @CsvSource(
        "'day-20-test', '16'",
        "'day-20', '0'"
    )
    fun testPartTwo(testFile: String, expectedResult: Int) {
        assertEquals(expectedResult, subject.partTwo(testFile))
    }
}
